Changing the base namespace
